% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackageone-package.R
\docType{package}
\name{testpackageone-package}
\alias{testpackageone}
\alias{testpackageone-package}
\title{testpackageone: My Package as a Demo}
\description{
This is my first package.
}
\author{
\strong{Maintainer}: Ian Proctor \email{ip6581a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
